+++ models/forms/deleteboard.js
@@ -26,6 +26,7 @@ module.exports = async (uri, board) => {
 		remove(`${uploadDirectory}/json/${uri}/`), //json
 		remove(`${uploadDirectory}/banner/${uri}/`), //banners
 		remove(`${uploadDirectory}/flag/${uri}/`), //flags
+		remove(`${uploadDirectory}/asset/${uri}/`), //flags
 	]);
 
 }