+++ CHANGELOG.md
@@ -33,6 +33,7 @@
   - Country blocklist now can actually fit all countries
   - Make "auth level" text box into "account type" dropdown in accounts page, easier to understand
   - Board owners can now edit custom pages
+  - Board owners can now manage custom assets
   - Show a little message and disable reply form on full threads (hit reply limit)
   - Allow longer language names for code blocks
   - User settings import and export option